// backend/src/types/independentObject.types.d.ts
export type IndependentObjectData = {
    documentNumber?: string;
    issueDate?: string;
    identifier: string;
    address: string;
    floor?: string;
    buildingIdentifier?: string;
    buildingPurpose?: string;
    buildingFloors?: string;
    landIdentifier?: string;
    objectPurpose?: string;
    levels?: string;
    area?: string;
    adjoiningParts?: string;
    sameFloorNeighbours?: string[];
    belowNeighbour?: string;
    aboveNeighbour?: string;
    previousIdentifier?: string;
};
//# sourceMappingURL=independentObject.types.d.ts.map